# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing data on oncological markers.",
        "title": "OncomarkerAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/unit-measure": {
            "get": {
                "description": "Retrieves a list of unit measure entries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UnitMeasure"
                ],
                "summary": "Get unit measure list",
                "responses": {
                    "200": {
                        "description": "Unit measure list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UnitMeasure"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing unit measure entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UnitMeasure"
                ],
                "summary": "Update unit measure",
                "parameters": [
                    {
                        "description": "Unit measure data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UnitMeasure"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated unit measure data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new unit measure entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UnitMeasure"
                ],
                "summary": "Create unit measure",
                "parameters": [
                    {
                        "description": "Unit measure data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UnitMeasure"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created unit measure data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/unit-measure/{id}": {
            "get": {
                "description": "Retrieves a unit measure entry by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UnitMeasure"
                ],
                "summary": "Get unit measure by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit measure ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit measure data",
                        "schema": {
                            "$ref": "#/definitions/model.UnitMeasure"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a unit measure entry by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UnitMeasure"
                ],
                "summary": "Delete unit measure",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Unit measure ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted unit measure data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.UnitMeasure": {
            "type": "object",
            "properties": {
                "full-text": {
                    "type": "string"
                },
                "global": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "shorthand": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}